put component to separate directory
